Extract AD auth to a new auth module - create space for NTLM
